fix: apply serving verdict in doctor path (01M1DGHKW5M2RQXPG5JDG66JVP)
